pyproject.toml
[project]
name = "stock-market-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "metaflow>=2.17.3",
    "polars>=1.32.3",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.7",
    "yfinance>=0.2.65",
]
[[project.authors]]
name = "scbtm"
email = "128545782+scbtm@users.noreply.github.com"

[build-system]
requires = [ "uv_build>=0.8.13,<0.9.0",]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "coverage[toml]>=7.10.5",
    "pre-commit>=4.3.0",
    "pyright>=1.1.404",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.10",
]

[project.scripts]
stock-market-analytics = "stock_market_analytics.main:main"
